LPS-56225 Add subpackage named 'test' in each package containing test classes and moved tests for it.
