+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/test/integration/com/liferay/dynamic/data/mapping/lar/test/DDMTemplateStagedModelDataHandlerTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.lar;
+package com.liferay.dynamic.data.mapping.lar.test;
 
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.TransactionalTestRule;
