+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/test/integration/com/liferay/dynamic/data/mapping/service/test/DDMTemplateServiceTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.service;
+package com.liferay.dynamic.data.mapping.service.test;
 
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.template.TemplateConstants;
