+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/test/integration/com/liferay/dynamic/data/mapping/storage/test/StorageAdapterTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.storage;
+package com.liferay.dynamic.data.mapping.storage.test;
 
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONSerializer;
